requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-discoveryengine>=0.17.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-genai==1.67.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.19
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.4.3
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
anthropic
